% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CollapseToLongestTranscript}
\alias{CollapseToLongestTranscript}
\title{CollapseToLongestTranscript}
\usage{
CollapseToLongestTranscript(ranges)
}
\arguments{
\item{ranges}{Signac annotation}
}
\value{
A grange object of gene coordinates
}
\description{
Obtain gene coordinates, a duplicate of Signac:::CollapseToLongestTranscript from Signac 1.14.0
}
