% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DistanceToTSS}
\alias{DistanceToTSS}
\title{Find peaks near genes}
\usage{
DistanceToTSS(peaks, genes, distance = 2e+05, sep = c("-", "-"))
}
\arguments{
\item{peaks}{A GRanges object containing peak coordinates}

\item{genes}{A GRanges object containing gene coordinates}

\item{distance}{Distance threshold. Peaks within this distance from the gene
will be recorded.}

\item{sep}{Separator for peak names when creating results matrix}
}
\value{
Returns a sparse matrix
}
\description{
Find peaks that are within a given distance threshold to each gene
a duplicate of Signac:::DistanceToTSS from Signac 1.14.0
}
