% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_with_assay}
\alias{validate_with_assay}
\title{validate_with_assay}
\usage{
validate_with_assay(res_df, sig_assay_gr, pvar = "padj", p_cutoff = 0.2)
}
\arguments{
\item{res_df}{A data frame of peak-gene associations results, containing columns of \code{gene}, \code{peak}, and pvar (see below).}

\item{sig_assay_gr}{A GRanges object for associated peak-gene pairs from other orthogonal assays, containing \code{gene} as a column in metadata}

\item{pvar}{Which variable in \code{res_df} to use for claiming significant associations. Default to 'padj'.}

\item{p_cutoff}{Cutoff for significance based on \code{pvar}. Default to 0.2.}
}
\value{
A named numeric vector that summarizes the overlap between \code{res_df} results and \code{sig_assay_gr} results
\describe{
\item{pval}{p value from one-sided Fisher exact test}
\item{n_overlap}{numebr of overlapped peak-gene pairs}
\item{oddsratio}{odds ratio in the two-by-two table}
\item{ratio}{overlapped pairs among all significant pairs from \code{res_df}}
\item{enr}{enrichment of overlapped pairs among significant pairs compared to all background pairs}
}
}
\description{
Validate peak-gene associations computed with single-cell multimodal data
by evaluating the consistency with orthogonal assay.
}
